feat(project): refactor fishes emulation & correct display

+++ server/src/parse_cfg.h
@@ -4,7 +4,7 @@
 struct config_t {
     int controller_port;
     int display_timeout_value;
-    int fish_update_interval;
+    float fish_update_interval;
 };
 
 int config_from_file(struct config_t *config, const char *cfgpath);
